--- .github/workflows/development.yml
@@ -1,59 +1,0 @@
-name: dev-static-website
-
-on:
-  push:
-    branches:
-      - '**'
-      - '!main'
-
-      
-jobs:
-  deploy:
-    name: Deploy to Development
-    runs-on: ubuntu-latest
-
-    steps:
-    - name: Checkout
-      uses: actions/checkout@v6
-
-    - name: Configure AWS credentials
-      uses: aws-actions/configure-aws-credentials@v4
-      with:
-        aws-access-key-id: ${{ secrets.MYKEY }}
-        aws-secret-access-key: ${{ secrets.MYSECRET }}
-        aws-region: ap-southeast-2
-
-    - name: Install pnpm
-      uses: pnpm/action-setup@v6
-      with:
-        version: 8
-        run_install: false
-
-    - name: Get pnpm store directory
-      shell: bash
-      run: |
-        #exit 1. # fail this workflow on purpose to test the retry workflow
-        echo "STORE_PATH=$(pnpm store path --silent)" >> $GITHUB_ENV
-
-    - uses: actions/cache@v5
-      name: Setup pnpm cache
-      with:
-        path: ${{ env.STORE_PATH }}
-        key: ${{ runner.os }}-pnpm-store-${{ hashFiles('**/pnpm-lock.yaml') }}
-        restore-keys: |
-          ${{ runner.os }}-pnpm-store-        
-
-    - name: Build and Test.
-      run: |
-        pnpm install
-        pnpm run test
-        pnpm run build
-
-    - name: Copy to dev.craiggunson.com bucket.
-      run: |
-        cd build
-        aws s3 sync . s3://dev.craiggunson.com --acl public-read --cache-control "max-age=60"  --delete \
-        --exclude '.git*' \
-        --exclude '.github*'
-        
-
