--- .github/workflows/production.yml
@@ -1,26 +1,0 @@
-name: static-website
-
-on:
-  push:
-    branches: 
-      - main
-
-jobs:
-  deploy:
-    name: Deploy Homepage
-    runs-on: ubuntu-latest
-
-    steps:
-    - name: Checkout
-      uses: actions/checkout@v6
-
-    - name: Configure AWS credentials
-      uses: aws-actions/configure-aws-credentials@v4
-      with:
-        aws-access-key-id: ${{ secrets.MYKEY }}
-        aws-secret-access-key: ${{ secrets.MYSECRET }}
-        aws-region: ap-southeast-2
-          
-    - name: Copy to craiggunson.com bucket.
-      run: |
-        aws s3 sync s3://dev.craiggunson.com s3://craiggunson.com --acl public-read --cache-control "max-age=300" --delete
