--- .github/workflows/retry.yaml
@@ -1,32 +1,0 @@
-name: retry
-on:
-  workflow_run:
-    workflows: ["dev-static-website"]
-    types:
-      - completed
-    branches:
-      - "**"
-
-defaults:
-  run:
-    shell: bash
-
-jobs:
-  retry:
-    runs-on: ubuntu-latest
-    permissions:
-      actions: write # Retry actions
-      checks: read   # Get info about the run
-    steps:
-      - name: retry
-        if: ${{ github.event.workflow_run.conclusion == 'failure' && github.event.workflow_run.run_attempt == 1 }}
-        run: |
-              output=$(gh run view --repo "${{ github.event.repository.full_name }}" "${{ github.event.workflow_run.id }}")
-              echo "$output"
-
-              #if echo "$output" | grep -q "The self-hosted runner lost communication with the server"; then
-              gh run rerun "${{ github.event.workflow_run.id }}" --repo "${{ github.event.repository.full_name }}" --failed
-              #fi
-        env:
-          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-
