+++ NotoEmoji-VariableFont_wght.ttf
