+++ index.html
@@ -1,0 +1,88 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="utf-8" />
+  <link rel="icon" href="kiwi.png" />
+  <meta name="viewport" content="width=device-width, initial-scale=1" />
+  <meta name="apple-mobile-web-app-capable" content="yes" />
+  <meta name="apple-mobile-web-app-status-bar-style" content="black-translucent" />
+  <meta name="theme-color" content="rgb(200, 100, 50)" />
+  <script src="https://kit.fontawesome.com/5239f6ce9b.js" crossorigin="anonymous"></script>
+  <meta name="description" content="Craig Gunson, Cloud and DevOps Tradie" />
+  <link rel="manifest" href="manifest.json" />
+  <title>Craig Gunson</title>
+  <link rel="stylesheet" href="style.css" />
+</head>
+<body>
+
+  <canvas id="canvas"></canvas>
+
+  <main>
+    <div class="section no-pad-bot" id="index-banner">
+      <div class="container">
+        <h1 class="center">Craig Gunson</h1>
+        <div class="row center">
+          <h2 class="col s12 light">Doing what I can, with what I have, where I am.</h2>
+        </div>
+      </div>
+    </div>
+
+    <div class="container">
+      <div class="section">
+        <div class="row">
+
+          <div class="col s12 m4">
+            <div class="icon-block">
+              <h2 class="center white-text"><i class="fa-solid fa-cloud-bolt fa-2x"></i></h2>
+              <h3 class="center">Cloud</h3>
+              <p class="center light white-text">Years of experience with Amazon Web Services and Google Cloud Platform, catering to their strengths to deliver customer value.</p>
+            </div>
+          </div>
+
+          <div class="col s12 m4">
+            <div class="icon-block">
+              <h2 class="center white-text"><i class="fa-solid fa-code fa-2x"></i></h2>
+              <h3 class="center">DevOps</h3>
+              <p class="center light white-text">Bringing software engineers, infrastructure, security and testing together, to swiftly deliver product.</p>
+            </div>
+          </div>
+
+          <div class="col s12 m4">
+            <div class="icon-block">
+              <h2 class="center white-text"><i class="fa-solid fa-seedling fa-2x"></i></h2>
+              <h3 class="center">Innovator</h3>
+              <p class="center light white-text">Moving at pace with Continuous Integration, Continuous Delivery, and AI.</p>
+            </div>
+          </div>
+
+        </div>
+      </div>
+    </div>
+  </main>
+
+  <footer class="page-footer">
+    <div class="container">
+      <div class="row">
+
+        <div class="center">
+          <h4>About</h4>
+          <p class="center white-text">I am passionate about making technology work for people.<br />
+            I spend my weekends outdoors, and weekdays on the computer.<br />
+            If you are curious, say hello.</p><br />
+            <a href="https://github.com/craiggunson"><i class="fab fa-github fa-3x" title="Github"></i></a>
+          <a href="https://www.linkedin.com/in/craiggunson/"><i class="fab fa-linkedin-in fa-3x" title="Linkedin"></i></a>
+          <a href="https://www.imdb.com/name/nm6043920/"><i class="fab fa-imdb fa-3x" title="IMDB"></i></a>
+        </div>
+
+
+      </div>
+    </div>
+
+    <div class="footer-copyright">
+      <div class="container">Aww Yeah!!!</div>
+    </div>
+  </footer>
+
+  <script src="main.js"></script>
+</body>
+</html>
