--- public/manifest.json
@@ -1,16 +1,0 @@
-{
-  "short_name": "homepage",
-  "name": "craiggunson.com",
-  "icons": [
-   
-    {
-      "src": "kiwi.png",
-      "type": "image/png",
-      "sizes": "120x120"
-    }
-  ],
-  "start_url": ".",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
-}
