+++ style.css
@@ -1,0 +1,193 @@
+@font-face {
+  font-family: Lato;
+  font-style: normal;
+  font-weight: 400;
+  src: local("Lato Regular"), local("Lato-Regular"),
+       url(S6uyw4BMUTPHjx4wXiWtFCc.woff2) format("woff2");
+  unicode-range: U+0000-00FF, U+0131, U+0152-0153, U+02BB-02BC, U+02C6,
+    U+02DA, U+02DC, U+2000-206F, U+2074, U+20AC, U+2122, U+2191, U+2193,
+    U+2212, U+2215, U+FEFF, U+FFFD;
+}
+
+html, body {
+  height: 100%;
+  width: 100%;
+  margin: 0;
+  padding: 0;
+  overflow-y: scroll;
+  overflow-x: hidden;
+  line-height: 1.5;
+  font-family: Lato, sans-serif;
+  font-weight: 400;
+  font-size: 3vh;
+  color: white;
+  min-height: 100vh;
+  background-repeat: no-repeat;
+  background-size: cover;
+}
+
+body::-webkit-scrollbar {
+  display: none;
+}
+
+canvas {
+  width: 100%;
+  height: 100vh;
+  display: block;
+  position: fixed;
+  top: 0;
+  left: 0;
+  z-index: -1;
+  pointer-events: none;
+}
+
+h1 {
+  color: white;
+  text-align: center;
+  line-height: 1.1;
+  padding: 0;
+  margin: 0;
+  font-size: 3em;
+}
+
+h2 {
+  font-weight: 400;
+  line-height: 110%;
+  margin: 1.78rem 0 1.424rem 0;
+  color: rgba(255, 255, 255, 0.9);
+}
+
+h3 {
+  font-weight: 400;
+  line-height: 1;
+  color: rgba(255, 255, 255, 0.9);
+}
+
+a {
+  text-decoration: none;
+  color: white;
+}
+
+i {
+  line-height: inherit;
+  letter-spacing: 1vw;
+  margin-bottom: 1vh;
+}
+
+.white-text {
+  color: white;
+}
+
+.center {
+  text-align: center;
+}
+
+.light {
+  font-weight: 300;
+}
+
+.icon-block {
+  padding: 0 2vw;
+}
+
+/* Layout */
+.container {
+  margin: 0 auto;
+  max-width: 100vw;
+  width: 95vw;
+}
+
+.row {
+  margin-left: auto;
+  margin-right: auto;
+}
+
+.row::after {
+  content: "";
+  display: table;
+  clear: both;
+}
+
+.container .row {
+  margin-left: -0.75rem;
+  margin-right: -0.75rem;
+}
+
+.row .col {
+  float: left;
+  box-sizing: border-box;
+  padding: 0 0.75rem;
+  min-height: 1vh;
+}
+
+.row .col.s12 {
+  width: 100%;
+  margin-left: auto;
+  left: auto;
+  right: auto;
+}
+
+@media (orientation: landscape) {
+  .row .col.m4 {
+    width: 33.3333333333%;
+    margin-left: auto;
+    font-size: 2vw;
+    left: auto;
+    right: auto;
+  }
+
+  .row .col.l6 {
+    width: 50%;
+    margin-left: auto;
+    left: auto;
+    right: auto;
+  }
+
+  .row .col.l3 {
+    width: 25%;
+    margin-left: auto;
+    left: auto;
+    right: auto;
+  }
+}
+
+.section {
+  padding-top: 1rem;
+  padding-bottom: 1rem;
+}
+
+.section.no-pad-bot {
+  padding-bottom: 0;
+}
+
+/* Footer */
+footer {
+  display: block;
+  position: static;
+}
+
+.page-footer {
+  text-align: center;
+  padding-top: 1vh;
+  background-color: rgba(0, 0, 0, 0.7);
+}
+
+.page-footer .footer-copyright {
+  min-height: 5vh;
+  text-align: center;
+  padding: 1vw 1vh;
+  background-color: rgba(0, 0, 0, 0.7);
+  font-weight: 300;
+  padding-bottom: 5vh;
+}
+
+/* Projects */
+.projecticon {
+  font-size: 1em;
+  font-weight: 401;
+  cursor: pointer;
+  color: white;
+  user-select: none;
+  -webkit-user-select: none;
+  -moz-user-select: none;
+}
